refactor(expense): migrate pay type and expense type to user dict

+++ src/main/java/top/aiolife/record/pojo/entity/IncomeEntity.java
@@ -43,6 +43,6 @@ public class IncomeEntity {
     /**
      * 收入类型ID
      */
-    private Integer incTypeId;
+    private Long incTypeId;
     private BigDecimal tax;
 }
